fix(gateway): omit reasoning from webchat transcript text
