+++ src/gateway/server-methods/webchat-transcript-reply-text.ts
@@ -1,0 +1,46 @@
+import { resolveSendableOutboundReplyParts } from "openclaw/plugin-sdk/reply-payload";
+import type { ReplyPayload } from "../../auto-reply/reply-payload.js";
+import { isAudioFileName } from "../../media/mime.js";
+import { sanitizeReplyDirectiveId } from "../../utils/directive-tags.js";
+import { isSuppressedControlReplyText } from "../control-reply-text.js";
+
+/**
+ * Synthesize plain-text for transcript rows from delivered reply payloads.
+ * Reasoning/thinking payloads must not contribute: they are shown as structured
+ * blocks on the main assistant turn (see #71910).
+ */
+export function buildTranscriptReplyTextFromPayloads(payloads: ReplyPayload[]): string {
+  const chunks = payloads
+    .map((payload) => {
+      if (payload.isReasoning) {
+        return "";
+      }
+      const parts = resolveSendableOutboundReplyParts(payload);
+      const lines: string[] = [];
+      const replyToId = sanitizeReplyDirectiveId(payload.replyToId);
+      if (replyToId) {
+        lines.push(`[[reply_to:${replyToId}]]`);
+      } else if (payload.replyToCurrent) {
+        lines.push("[[reply_to_current]]");
+      }
+      const text = payload.text?.trim();
+      if (text && !isSuppressedControlReplyText(text)) {
+        lines.push(text);
+      }
+      for (const mediaUrl of parts.mediaUrls) {
+        if (payload.sensitiveMedia === true) {
+          continue;
+        }
+        const trimmed = mediaUrl.trim();
+        if (trimmed) {
+          lines.push(`MEDIA:${trimmed}`);
+        }
+      }
+      if (payload.audioAsVoice && parts.mediaUrls.some((mediaUrl) => isAudioFileName(mediaUrl))) {
+        lines.push("[[audio_as_voice]]");
+      }
+      return lines.join("\n").trim();
+    })
+    .filter(Boolean);
+  return chunks.join("\n\n").trim();
+}
